package_tracker_pl/app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
paho-mqtt==2.1.0
